flutter analyze --watch check repo packages for conflicts
